tests: Move all docker_registry_image build tests to docker_image.
