+++ testdata/resources/docker_image/testBuildDockerImageKeepConfig.tf
@@ -1,0 +1,14 @@
+provider "docker" {
+  alias = "private"
+}
+resource "docker_image" "foo" {
+  provider             = "docker.private"
+  name                 = "%s"
+
+  build {
+    context      = "%s"
+    remove       = true
+    force_remove = true
+    no_cache     = true
+  }
+}
