+++ testdata/resources/docker_image/testBuildDockerImageNoKeepJustCache.tf
@@ -1,17 +1,14 @@
 provider "docker" {
   alias = "private"
-  registry_auth {
-    address = "%s"
-  }
 }
-
-resource "docker_registry_image" "file_permissions" {
+resource "docker_image" "%s" {
   provider             = "docker.private"
   name                 = "%s"
-  insecure_skip_verify = true
 
   build {
     context      = "%s"
-    dockerfile   = "%s"
+    remove       = false
+    force_remove = false
+    no_cache     = false
   }
 }
